EXTJS-26715: Change folder structure for SplitButton
