+++ packages/reactor-kitchensink/src/examples/SplitButton/SplitButton.js
